Setup base layout
Prepare COS_ACTIVE and COS_PASSIVE for A/B upgrades, COS_PERSISTENT for /usr/local and COS_OEM for /oem to store default cloud-init files.
Also make the init to load those partition on boot, and setup grub failback partitions as well in the installer. It also adds the possibility to supply
a default config cloud-init file which is downloaded during the installation process. At the moment requires curl, which is broken due to the fact that depends on perl.

It renames 'installer' to 'cos-installer'

It also add cos-upgrade to handle the upgrade process in the A/B Dual copy fashion, but currently it fails due to the partition size which aren't sufficient to pull down the new image.

It needs also setup of /etc/ssl, which is now absent and thus prevents to pull images from docker registries ( needs to create /etc/ssl properly, and call `update-ca-certificates`)

Adds also device mapper to initramfs

Relates to #5, #8 and #37