Enlarge ACTIVE/PASSIVE partition sizes
Now they defaults to 2GB.
It now also detect booting partition in cos-upgrade, doing the partition switch based on the one we are booting into

Also, setup TMPDIR and XDG_RUNTIME_DIR until we have a persistent state partition for the update process.
TMPDIR for luet can be annotated in the config file, but temporarly do it here until the process is completely fixed

Also wipe passive partition, until we get a persistent db for luet.

Add gawk to the base image, which is used in cos-upgrade

Related to #8