Manipulate /etc/issue after install and run cos-upgrade from livecd
The default cloud-init file, serves to the purpose of illustrating how it works, and at the same
time informs on how to run upgrades.

Besides, let the user choose if cos-upgrade have to be run on the found
partitions and assume we are about to change COS_ACTIVE. Ask the user
when no specific options are given.

In this way, cos-upgrade can also be used from the LiveCD to recover the
ACTIVE partition automatically.