Update grub during upgrade
This is needed because as we don't have a peristent state (yet) in any
device, we wipe and treat it as a fresh install.

The grub needs to be re-generated as it is part of the installation
process. Now that is bundled in system/cos we don't have to generate it
on the fly.