Require /oem mount point for cos-setup-fs.service
This commit adds a dependency of cos-setup-fs.service with the /oem mount
point. In addition the timeout of the /oem and /usr/local is raised a
little so it is really unlikely they are missed on the first attempt to
mount them and their dependency to /etc is dropped. In fact we do not
care about what is mounted first /etc or /oem, we care that both are
mounted when cos-setup-fs.service is executed.