Add a Fedora cOS variant
This commit adds a variant for cOS based on Fedora. Small refactor in
Makefile has been done to allow different flavors. Each flavor maps to a
set to an specific yaml file including the required additional values.
The default flavor is set to opensuse.

In addition the kernel package has been dropped and included as part of
the base distro package, this is done to facilitate templating across
different distros.

Finally there is an unrelated update for the mount point of the ephemeral
overlay volume that gets mounted in /run/overlay now instead of
/data.overlay.

Fixes #61