Add cos-feature
Allows to enable/disable default profiles in a cOS system.
Some default profiles are provided as is: vagrant, k3s, harvester,
carrier.

Multiple arguments are supported. This allows also to tweak the packer
templates to enable different features and making possible to bake
different flavors with the CI.

It does also a bunch of changes in the initrd:

- Add rmdir to initrd
- Omit multipath in initrd:
  multipath clashes with isoscan, preventing system to boot correctly

Signed-off-by: Ettore Di Giacinto <edigiacinto@suse.com>