Embed grub configuration into the image
This commit embedds the grub configuration into the image, in fact, the
embedded part affects only the boot arguments which are provided within
the image.

The main grub.cfg is still static but provided as an static file within
the image too.

This commit allows us to change boot parameters on upgrades and also
easily allows the user to include additional parameters at run time if
needed.

Fixes #109

Signed-off-by: David Cassany <dcassany@suse.com>