Add log capture to test facility
Now in order to capture any log files from the test machines we have no
way of doing that.

This pathc introduces a method to be able to hook on AfterEach test that
will try to gather logs from the remote machine to the local one.
The method will not fail if there is no log files in the remote machine
to avoid blocking a valid job.

Also adds the facility to the upgrade jobs to gather the image mtree
logs and adds to the tests on CI the store rtifacts step to stores those
captured logs, currently only fired for upgrade tests

Signed-off-by: Itxaka <igarcia@suse.com>