Do not check mtree for repository/tree files
Now that we are emmiting events for image unpack events, when doing a
normal cos-upgrade the repository adn tree files will be downloaded and
extracted. This collides with our mtree check script as those files do
not contain mtree values but our script still tries to check them.

This patch adds exceptions to those events by matching the image tag

Signed-off-by: Itxaka <igarcia@suse.com>