[mtree] rework the mtree check script
As we now support unpack events, both paths (channel upgrades or single
docker image) send the same events now, so we can hijack that to have a
single path with fewer lines of code.

This patch simplifies the script to have one singular path that both
upgrade methods use and verify the mtree checksums. It also cleans up a
bit the logging from the script to be more useful.

Signed-off-by: Itxaka <igarcia@suse.com>