Add image verification for single images (#234)
* Add image verification for single images

When upgrading by passing a single docker image we need to also
verifythat the contents from the image have the same checksum as
the one stored in the metadata

This expands the existing image-mtree-check script to take into
consideration the unpack event and act accordingly

Signed-off-by: Itxaka <igarcia@suse.com>

* Fix wrong usage message on cos-upgrade

Signed-off-by: Itxaka <igarcia@suse.com>

* Add exit status to mtree check on single docker image

Signed-off-by: Itxaka <igarcia@suse.com>

* Use unpack plugin events for mtree checks

On luet 0.16.5 luet now supports emmiting events when running the unpack
action which we can use to verify the mtree checks during single image
upgrade

Signed-off-by: Itxaka <igarcia@suse.com>

* Reenable more test related to signed images

Now that verify is in for all types of upgrade, re enable some of the
tests that we dropped before

Signed-off-by: Itxaka <igarcia@suse.com>

* Add log capture to test facility

Now in order to capture any log files from the test machines we have no
way of doing that.

This pathc introduces a method to be able to hook on AfterEach test that
will try to gather logs from the remote machine to the local one.
The method will not fail if there is no log files in the remote machine
to avoid blocking a valid job.

Also adds the facility to the upgrade jobs to gather the image mtree
logs and adds to the tests on CI the store rtifacts step to stores those
captured logs, currently only fired for upgrade tests

Signed-off-by: Itxaka <igarcia@suse.com>

* [vagrant] forward ssh port on libvirt

Until very recent versions (20 days ago) libvirt did not enable the ssh
port forwarding by default arguing that you can already access the
machine directly via other methods.

For local testing thougth, its needed so tests can run and connect to
the machine.

This patch explicitely enables ssh forwarding for libvirt machines

Signed-off-by: Itxaka <igarcia@suse.com>

* Do not check mtree for repository/tree files

Now that we are emmiting events for image unpack events, when doing a
normal cos-upgrade the repository adn tree files will be downloaded and
extracted. This collides with our mtree check script as those files do
not contain mtree values but our script still tries to check them.

This patch adds exceptions to those events by matching the image tag

Signed-off-by: Itxaka <igarcia@suse.com>

* [vagrant] bump libvirt specs

Default libvirt specs are 512Mb of ram and 1 vcpu.

This patch bumps that to fall in line with the vbox configuration fo
4096Mb of ram and 2 vcpus

Signed-off-by: Itxaka <igarcia@suse.com>

* [mtree] rework the mtree check script

As we now support unpack events, both paths (channel upgrades or single
docker image) send the same events now, so we can hijack that to have a
single path with fewer lines of code.

This patch simplifies the script to have one singular path that both
upgrade methods use and verify the mtree checksums. It also cleans up a
bit the logging from the script to be more useful.

Signed-off-by: Itxaka <igarcia@suse.com>

* [tests] Skip capturing logs if test didnt fail

Instead of trying to capture the logs for mtree on every test just do it
if the test fails, so we can save some time, otherwise we are just
waiting for the scp to timeout.

Signed-off-by: Itxaka <igarcia@suse.com>

* [tests] Expand on upgrade tests

Fix one test that was using a broken image and in turn use that broken
image to test for failure when verifying.

Signed-off-by: Itxaka <igarcia@suse.com>