Use the new luet-mtree event for upgrade checks
We were using a bash script to call luet-mtree on upgrade to check for
valid mtree checksums.

luet-mtree 0.0.6 now supports working with events from luet itself,
which will do the same as the bash script and extract the image metadata
to check the mtree checksums.

This patch removes the old bash script and uses the luet-mtree plugin
directly to check the checksums, it also bumps luet-mtree to 0.0.6

Signed-off-by: Itxaka <igarcia@suse.com>