Add SLE distribution
This distribution is functionally equivalent to opensuse with the key
difference being that the only gpg keys trusted are from SUSE build
service.  This ensures no community built binaries are in the final
build.

Since this distribution is basically the same as opensuse I added
the field "distribution_like" to the values to mimic the behavior of
ID_LIKE in os-release.

Signed-off-by: Darren Shepherd <darren@rancher.com>