Improve cloud-config (#361)
This commit moves the recovery_mode detection to the very very first
stage where 'yip' comes into play. This is specially helpful to make
use of it in later stages. In addition the default layout is changed for
recovery and system partitions, in recovery mode the COS_PERSISTENT
partition is ignored at boot time. This is is specially helpful for
vanilla images, with the change they can boo the recovery system as
they are, without requiring any partitioning at first boot.

Signed-off-by: David Cassany <dcassany@suse.com>