[ci] capture serial log on tests
Sometimes having the serial log output can help us troubleshoot failures
in CI.

This patch enables the serial output to a file during tests and image
creation with vbox and stores it as an artifact as part of the PR
workflow.

Unfortunately as a side effect we cannot longer use the snapshot restore
method after each test to restore the test vm to a pristine state as it
will truncate the file, losing all logs. This patch just reverts to the
old reset method.

Signed-off-by: Itxaka <igarcia@suse.com>