Add /etc/hosts file back
While it fixes things for from scratch installs, it breaks many things:

- we should generate an /etc/hosts file during boot, ideally in
  cloud-init phase, but that looks cumbersome as we can just relay on
what's the default
- going to that route, and then excluding /proc, /sys, /run and /tmp
  requires further work on the bootstrap part, currently tests are not
passing (see #366)

It sounds easier instead to have a separate package for it just for
creating derivatives "from scratch" - in that way we don't interfere
with the standard system/cos which is supposed to run on ISO and other
medium too, where such folders are mandatory.

This addition adds a system/cos-container to just provide a separate package
with /proc, /run, /sys, /dev, /tmp stripped which is meant to be used for
from scratch scenarios.

Supersedes: #366

Signed-off-by: Ettore Di Giacinto <edigiacinto@suse.com>