Make use of grub2 environment block
This commit sets the grub2 configuration to load /grubenv file, if any,
and checks the `next_entry` variable to set the default grub2 option.

This is helpful to set default boot in grub only for the next boot. From
cOS it is enough to run `grub2-editenv /oem/grubenv set
next_entry=recovery` to set the next reboot to default to recovery system.

This commit also adds a new docs page devoted to grub configuration.
Finally in also moves the wiki page devoted to a K3s and Fleet
deployment example into its own docs page.

Signed-off-by: David Cassany <dcassany@suse.com>