amazon: Enhance AMI name and tags
This patch adds the github.sha to the build name of the AMI so we can
easily identify where the image came from if pushed from CI and to be
able to upload more than one image per day on CI as well.

Also adds it to the tags, and add a new Flavor tag, so we can use data
lifecycle policies to auto remove images and search for an X flavor by
using those tags.

Also adds that tag to the aws_upload.sh script to tag the recovery
images as "recovery"

The default value for no git_sha is none.

Signed-off-by: Itxaka <igarcia@suse.com>