Fix 481 (#509)
* ci: Use the insecure vagrant key for testing

Using user/pass its cool but you know whats even cooler? using the key
that we insert as a vagrant cos feature which persists across reboots
and resets

Signed-off-by: Itxaka <igarcia@suse.com>

* Fix upgrading script using privileged to extract

Looks like becasue the dir we point unpack to doesnt exists **and** its
has a parent that also doesnt exist the non privileged path on luet[0]
just tries to create the parent but fails because its not recursive,
i.e. doesnt create the parents as neccesary. Meanwhile when using the
privileged path[1] the lib creates the destination and all the parents
needed, thus this is not a privs issue but a library issue. It only
appears when having that switch because it takes different paths to
different libs to deal with this.

Creating the expected dir before extracting works in both paths.

[0] https://github.com/mudler/luet/blob/master/vendor/github.com/containerd/containerd/archive/tar.go#L231
[1] https://github.com/mudler/luet/blob/master/vendor/github.com/docker/docker/pkg/archive/archive.go#L942

Signed-off-by: Itxaka <igarcia@suse.com>

* Bump cos

Signed-off-by: Itxaka <igarcia@suse.com>