ci: Add test for deploying from raw image
Recent events have shown that we do not test cos-deploy from a raw image
as well as we should.

This patch adds a simple test and preparation of a vm based of the raw
image so we can test the the default cos-deploy stanza works as expected
and lays the base so we can expand it with more tests in the future.

Unfortunately as this test is fully modifying the disk image we can only
test one thing per time, once we modify the disk the pristine state that
we started with is gone, so we should probably create one work per test
that we want.

This also means that we can drop the normal raw-image-test that only
checks if we can boot from the raw image properly as this new test already
checks that part.

Signed-off-by: Itxaka <igarcia@suse.com>