Uniform cos config file sourcing
Allow to use os-release and consistently source from /etc/cos/config whenever found.
In upgrade it moves the sourcing inside find_upgrade_channel as otherwise
the content in the config would override what it would be manually passed by.
This could cause suc-update to not work as expected as it would use the value
in the configs (if applied) instead of the ones passed by argument.

Signed-off-by: Ettore Di Giacinto <edigiacinto@suse.com>