ci: Add arch to CI builds
This adds an arch value to the ci templates in order to differentiate
between the current amd64 and future arches that we may add.

On top of that it expands the ci templates with the labels value which
would add an extra step to the workflow in order to check for the given
labels in the PR and fail the workflow otherwise (Useful for arches,
do-not-test, etc...)

It also adds an arm64 workflow that runs on self runners for PRs labeled
with the arm64 label. Currently a separated one as its experimental..and
we dont have arm54 runners yet :D

Signed-off-by: Itxaka <igarcia@suse.com>