ci: Add arch to CI builds (#614)
* ci: Add arch to CI builds

This adds an arch value to the ci templates in order to differentiate
between the current amd64 and future arches that we may add.

On top of that it expands the ci templates with the labels value which
would add an extra step to the workflow in order to check for the given
labels in the PR and fail the workflow otherwise (Useful for arches,
do-not-test, etc...)

It also adds an arm64 workflow that runs on self runners for PRs labeled
with the arm64 label. Currently a separated one as its experimental..and
we dont have arm54 runners yet :D

Signed-off-by: Itxaka <igarcia@suse.com>

* ghrunner: missing deps

Add missing deps to runner image (sudo, git) and fix the labels option
for the runner

Signed-off-by: Itxaka <igarcia@suse.com>

* ghrunner: bump runner version

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: introduce skip_docker_build

In case we want to skip the docker builds

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: disable iso/tests for arm64

Currently experimental so we better dont spend cycles on that

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: Set different repos if arch is not amd64

We agreed that the current amd64 repo will kep its name as its the
"default" or what we expect most users to use and for the rest of arches
we would use the suffix $arch on them

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: Small fix to skip images in github-release

if the images are skipped we need also to not have them as need in the
release-github job

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: add release for arm64

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: Fix image link to use the arch in the file name

Also fixes a small issue with github-release which was downloading all
files without checking if they were requested

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: rename arm files to arm64

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: add master for arm64

Adds master workflow for arm64 which will build and push all flavors on
arm64 branches

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: Allow not installing luet from cos repo

Currently we use the cos repo to install or "upgrade" luet but for newly
arches we may not have pushed the required artifacts yet. This allows a
different path by just copying the downloaded luet into its proper dir

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: dont install luet from repo for arm64

At least we need this to pass 1 master run to have the proper artifacts
in place

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: clean luet binaries and paths on local runners

Otherwise the install script will fail and we want a clean start

Signed-off-by: Itxaka <igarcia@suse.com>

* fixup: I dont know how to write the mv command

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: Fix wrong repo for luet

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: separate arch values and change amd64 name

Separate the values files into archs so it build the correct packages
(golang)

Also change from amd64 to x86_64

Signed-off-by: Itxaka <igarcia@suse.com>

* templates: introduce do_checksum template

Useful to auto check if we want the checksum or not

Signed-off-by: Itxaka <igarcia@suse.com>

* build.go: Allow for skipping packages to build

The automatic checker tries to get all missing packages but does not
accep a skip packages var in order to be able to skip some packages.

This introduces a new var SKIP_PACKAGES in the same format as the
PACKAGES var for building that accepts a space separated list of
packages in the format category/name and will skip those packages from
the automated build (which is the only thing using the build.go stuff I
think)

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: skip fips related packages in arm64

There is no arm64 (or other arches than x86_64) golang-fips available

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: disable building live/grub* packages

currently failing as they depend on the proper arch

Signed-off-by: Itxaka <igarcia@suse.com>

* ci: drop more live-* packages from arm64 builds

Signed-off-by: Itxaka <igarcia@suse.com>

* arm64: fix grub2-efi-image package

Now it supports the arm64 path to generate a valid package

Signed-off-by: Itxaka <igarcia@suse.com>

* fix get_luet script

Signed-off-by: Itxaka <igarcia@suse.com>

* k9s: add binutils to prelude

Seems like on arm64 binutils is not installed

Signed-off-by: Itxaka <igarcia@suse.com>

* makefile: Fix cache for luet based off the current arch

Signed-off-by: Itxaka <igarcia@suse.com>

* k9s: add binutils-gold package install

Looks like compiling k9s in arm64 requires the gold linker

Signed-off-by: Itxaka <igarcia@suse.com>

* values: drop i386 package from arm64 build

Signed-off-by: Itxaka <igarcia@suse.com>