ci: rework ci to split workflows by flavor+arch
This pathc introduces a new level on the config files for the workflows
that refers to the flavor being tested.

the templates are mostly unchanged as we already dealt with the
different flavors correctly, the only thing tht needed a bit of change
was the config file and those changes are mostly minimal thank to using
anchors.

Basically with this we split all the workflows by flavor+arch and allows
us to have more fine tuning of each job. With also the addition of
future switches (i.e. run_tests, release_github, etc...) we can have
much more control per job instead of having all the options for all
flavors in the same blob. As shown on this change, is now pretty clean
the differences between the green and blue job for example and what is
run on one vs the other.

As a side effect, if one job fails we no longer have ot retriggers the
full workflow with all the flavors, as that never made any sense to have
them together, but insteaed trigger each individually

Signed-off-by: Itxaka <igarcia@suse.com>