test: split vagrant in 2 machines
As we are only using vbox machine for x86_64 tests, it makes no sense to
have a different provider for the same machine.

Instead split it into 2 different machines with different names and
different configs for easy managing.

This patch also uses a var for the BOXNAME being used and creaates that
var based of the current arch.

The boxfile also takes into consideration the arch while looking for the
vagrant box

Signed-off-by: Itxaka <igarcia@suse.com>