test: Add a function to get teh ssh ip from vagrant for arm64
As we are using libvirt for running arm64 test instances and on reboot
they will change the IP, we just add a check to get the IP from vagrant
directly as it keeps updated in there.

Unfortunately any local forwarding set by vagrant gets lost on reboot so
we cannot use that directly.

Signed-off-by: Itxaka <igarcia@suse.com>