Make OEM partition label customizable at build time
This commit sets the basis to make OEM partition label customizable.
Currently it can only be customized a build time, it already was,
however the value was hardcoded in several places. Now the label
value is included in /etc/cos/config file and all other packages
read it from there at build time and install time.

In the future it should not be complex to make it configurable at
install time. For that some little extensions to the installer are
required.

Signed-off-by: David Cassany <dcassany@suse.com>