Bind mount initrd's tmpfs /tmp at initramfs stage
This commit bind mounts /tmp from initrd environment to initramfs stage
environment (chroot on /sysroot). This enables the use of tools or
utilites in initramfs stage that relay on a writable /tmp.

Note this is an ephemeral filesystem that gets lost at switching root
stage, so right after initramfs stage.

Signed-off-by: David Cassany <dcassany@suse.com>