Use latest luet from script in the publish job
github workers are symlinking /bin to /usr/bin/ which results in luet
identifying the luet-cosign plugin twice and calling it twice which
results in failure to sign the artifacts due both plugins accessing the
same files at the same time.

This makes it so luet is installed from the get_luet.sh script directly,
which would install the latest version, which has a fix to have unique
plugin names in the go-pluggable lib[0]

This commit is safe to revert once the luet and luet-cosign packages
have been published as there contain fixes for this issue.

[0] https://github.com/mudler/go-pluggable/commit/9263b05c562e6207999fd74bebf2cf36fa013e02

Signed-off-by: Itxaka <igarcia@suse.com>