Move elemental to git based tracking for versions
Instead of having to bump the version on elemental each time we need a
change, we can move to using a git_sha and keep our versions separated
from the elemental ones for now as we may need pathes and commits from
elemental and we dont want to keep generating releases.

Once elemental is stable we can change back to using the package version
as the elemental version again, thus for now we should only bump the
revision and keep the version number at 0.0.12 until then.

Signed-off-by: Itxaka <igarcia@suse.com>