Refactor CopyActive method to turn tin CopyImage (#95)
This commit refactors CopyActive method to become CopyImage, where
the image being copied is passed as a parameter. This makes it easy to
reuse it other contexts beyond copying the active image.

In the same scope the refactor includes the image source struct as part
of the v1.Image struct. So all the details of the image are
self contained.

Signed-off-by: David Cassany <dcassany@suse.com>

Co-authored-by: Itxaka <igarcia@suse.com>