Disable most of the EFI tests in VBox
This commit disables several EFI installer tests as they consume quite
some cycles of the CI infraestructure and they often fail due to
unstable EFI support in VBox. Because of this situation we are not
relaying on them for anything and failures are ignored.

With this change we expect to free some of the CI resources and stil
keep a basic EFI boot afer install.

Since there will be way less reboots and environment cleanups we expect
to also increase the test success rate, as failures were often
happening while restoring the environment and rebooting between tests.

Signed-off-by: David Cassany <dcassany@suse.com>