Add service that allows us to control the cdrom eject
This creates a service file that we can control by setting
/run/cos/ejectcd and that will also check to see if we booted with the
cdboot stanza in the cmdline in order to trigger.

By default is set to run after shutdown and reboot if those two
conditions met.

This allows us and derivatives to control ejecting the cd at the end of
an install for example

Signed-off-by: Itxaka <igarcia@suse.com>