Auto tag on cos/system bump
This patch will run on schedule daily and get the current version set
for cos/system, the latest tag of the repo and compare both. If there is
a discrepancy on them, it will use that new cos/system tag to create a
new tag on the repo, which should trigger the release jobs to create a
new release with that new tag.

Signed-off-by: Itxaka <igarcia@suse.com>