Fix tag and tag pushing for autorelease
Ignore any build number (i.e. -NUMBER) that the bot auto pushes when
comparing version as we only want the semver to be pushed.

Also push the tag ourselves with the rest api so we can use the CiBot
PAT, which means that release jobs will be triggered on tag release.
Otherwise, because the tag cames from the github actions itself, the
jobs wont trigger. This is a deliberate decision from github in order to
avoid circular jobs.

Signed-off-by: Itxaka <igarcia@suse.com>