Use /etc/os-release on install and reset (#218)
This commit makes `install` and `reset` commands to read the `/etc/os-release` file of the installed image to load the `GRUB_ENTRY_NAME` variable, if set, it is used as the default grub menu entry. Note this value can also be configured within the `config.yaml` file. The configuration file has higher priority.

In addition this commit also also adds the `grub-entry-name` configuration option on upgrade command, so `install`, `reset` and `upgrade` share the same behavior in order to set the grub menu entry.

Signed-off-by: David Cassany <dcassany@suse.com>