Drop COSIGN_REPOSITORY from our side (#1482)
There is no need to push to a different repo the signatures as now quay
supports the cosign signatures.

The value its still there as its picked by cosign, so in the end there
is still the possibility of using it when building a derivative for
example, but its not needed anymore on our side so lets drop it and
update the resigner as well.

Also drops 2 old jobs that were not working

Signed-off-by: Itxaka <igarcia@suse.com>