[ci] Re-enable build-pr-docker
Was disabled due to running out of space in the github provided runners.

Use the newer build.go facility to build, which requires a different
docker image with the binary on it and less stuff copied.

Drops the .envrc as it was mostly not being used

Updates the docs to reflect the new method used to build in a container

Signed-off-by: Itxaka <igarcia@suse.com>