Add extra partitions config on installation (#306)
* Add extra partitions config on installation

This adds a new key to the install config that allows to pass a list of
extra partitions to add during install.

Same keys as the usual partitions exists, name, label, fs, mountpoint,
etc..

Also allows to override the sizes and reorders the partition creation to
take that into account, as the partition with size == 0 (which means use
all the free space available) has to go the last.

This allows creating extra partitions with size 0 and will reorder the
partitions properly to account for this. Also forbids having several
partitions with size == 0 or a persistent and extra partition with both
having size == 0

Signed-off-by: Itxaka <igarcia@suse.com>

* Update config example

Signed-off-by: Itxaka <igarcia@suse.com>

* Lint

Signed-off-by: Itxaka <igarcia@suse.com>

* fix tests, remove focus

Signed-off-by: Itxaka <igarcia@suse.com>

* comments

Signed-off-by: Itxaka <igarcia@suse.com>

Signed-off-by: Itxaka <igarcia@suse.com>