Add shim to grub2-efi packages
Add the shim.efi to the grub2 packages (system adn live) in a way that
its backwards compatible.

As shim is just a launcher for grub.efi, it should be backwards
compatible with anything using this packages plus compatible with
elemental-cli versions that now use the shim for the installation as
well as older versions which did not do that.

Signed-off-by: Itxaka <igarcia@suse.com>