Fix signing referenceId
When pushing the artifacts and signing them we use 2 different
apporaches to getting the current referenceID/tag/commit.

This could result in pushing one reference and using a different
one for signing.

This patch fixes that by moving the signing to the makefile,
as it should have been from the start, and reusing the same
reference for both pushing the artifacts and signing them, so
there is no possibility of mixing them up

Signed-off-by: itxaka <igarcia@suse.com>