Fixes cos-setup systemd unit paths
'/lib/systemd/system' is not the path used for most major distributions.
In fact Fedora and Ubuntu ship their /lib as a symlink to /usr/lib,
hence the real path is /usr/lib/systemd/system. In openSUSE Leap there
is no such link and both exist /usr/lib/ and /lib, being /usr/lib the
default path for systemd. This also applies for Teal.

Signed-off-by: David Cassany <dcassany@suse.com>