Wait for all persistent volumes
This commit adds a wait logic for all persistent volumes configured
within the immutable-rootfs dracut module. It has timeout of 60 seconds
for each volume.

Note at this stage the root device is already found and mounted, since
all volumes are expected to be part of the same disk it is really
unlikely the wait is needed.

In addition this commit does not ignore any mountpoint in case of
errors. It keeps creating the fstab as expected if the mountpoint
could not be mounted or prepared.

Signed-off-by: David Cassany <dcassany@suse.com>