Use build-disk command
This commit changes the way we build our test disks.

It uses the new build-disk command instead of losetup and the install
command. This means we can run build-disk without privileges and could
also build expandable images.

As of now we build the full image and boot into active to run the tests
right away.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>