Prevent unnecessary copy of image tree to tmp folder
This commit prevents copying already accessible image
trees to a temporary foldery to later on sync it again
to the final filesystem image.

Prevents rsyncing twice the root tree for directory
and file image sources. This tends to be the case for
recovery image deployments, which usually are just a
copy of the local active system.

Signed-off-by: David Cassany <dcassany@suse.com>