make configurable boot-assessment checks (#2018)
* make configurable boot-assessment checks

Signed-off-by: David Cassany <dcassany@suse.com>

* Improve EventuallyConnects in tests

Signed-off-by: David Cassany <dcassany@suse.com>

* Fix fallback test

Signed-off-by: David Cassany <dcassany@suse.com>

* Store boot assessment variables in grubenv

This commit refactors the boot-assessment logic to run checks
with a customizable service and to store grub variables in
already existing files in EFI partition and OEM.

Also the EFI partition is made accessible in after-*-chroot
hooks. This makes easier to write and manage files in EFI
partition if needed (e.g. grub_oem_env variable file), without
having to relay on mounts or remounts.

Signed-off-by: David Cassany <dcassany@suse.com>

---------

Signed-off-by: David Cassany <dcassany@suse.com>