Add persistent disk encryption
Add flags to encrypt persistent partition on install:

* encrypt-persistent: flag to enable luks encryption on persistent
  partition.
* enroll-passphrase: string to enroll as passphrase to unlock partition.
* enroll-key-file: key-file to enroll as key to unlock partition.

During install this will invoke cryptsetup to create the LUKS partition
and during mount we use systemd-cryptsetup to attach the partition
before mounting the contained filesystem.

This also introduces some changes in the grub configuration, the
encrypted_volumes variable can be set in grub_oem_env during install to
configure which volumes are actually encrypted.

Using a config-file it is also possible to encrypt any extra-partitions
using the following syntax:

```yaml
install:
  extra-partitions:
    - Name: extra
      size: 100
      fs: ext4
      label: extra
      encryption:
        name: cr_extra
        key_slots:
        - slot: 1
          passphrase: "extrapass"
```

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>