SELinux relabel on boot
This commit introduces a new command 'elemental relabel' for relabeling
files and directories.

The command is invoked during initramfs stage in the new optional
'selinux' feature.

In essence it runs setfiles in mounted persistent and ephemeral
directories in order to circumvent a bug with overlayfs and
selinux copy-up on xattrs.

During mount we also put a list of persistent+ephemeral directories in
/run/systemd/extra-relabel.d/elemental.layout in order to make systemd
relabel the directories before loading the policy.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>