SELinux relabel on boot
This commit introduces a new command 'elemental relabel' for relabeling
files and directories.

The command is invoked during initramfs stage in the new optional
'selinux' feature.

In essence it writes all ephemeral and persistent directories to
/run/systemd/relabel-extra.d/elemental.layout and then runs setfiles in
those directories in order to circumvent a bug with overlayfs and
selinux copy-up on xattrs.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>