SELinux relabel on boot
In this commit we add selinux support to the mount command.

During mount we put a list of persistent+ephemeral directories in
/run/systemd/extra-relabel.d/elemental.layout in order to make systemd
relabel the directories before loading the policy.

We also try to chroot into the new sysroot and run setfiles using a find
wrapper to set on deepest files first.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>