Do not return error for efi.ReadLoadOption
When encountering a boot-entry for a device-path that go-efilib is
unable to parse we should just skip that entry instead of erroring out.

This same behavior is used earlier in the same loop so this should fix
an edge-case where the boot entry is parseable but the device-path is
not.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>